Read uint

+++ EsfLibrary/Esf/AbcbFileCodec.cs
@@ -9,7 +9,9 @@ namespace EsfLibrary
     {
         public AbcbFileCodec() : base(0xABCB) { }
         protected override string ReadUtf16NodeNames(BinaryReader reader) {
-            int strLength = reader.ReadInt32();
+            // I assume the format serializes as uint; as int for length does not make much sense
+            // In practice the infrastructure does not support values exceeding int.MaxValue anyway
+            int strLength = (int)reader.ReadUInt32();
             return Encoding.Unicode.GetString(reader.ReadBytes(strLength * 2));
         }
         protected override void WriteUtf16NodeNames(BinaryWriter writer, string toWrite) {
@@ -17,7 +19,7 @@ namespace EsfLibrary
             writer.Write(Encoding.Unicode.GetBytes(toWrite));
         }
         protected override string ReadAsciiNodeNames(BinaryReader reader) {
-            int strLength = reader.ReadInt32();
+            int strLength = (int)reader.ReadUInt32();
             return Encoding.ASCII.GetString(reader.ReadBytes(strLength));
         }
         protected override void WriteAsciiNodeNames(BinaryWriter writer, string toWrite) {
